+++ fuel-tx/src/transaction/validity/error.rs
@@ -2,7 +2,7 @@ use crate::UtxoId;
 use fuel_types::{
     AssetId,
     ContractId,
-    MessageId,
+    Nonce,
 };
 
 /// The error returned during the checking of the transaction's validity rules.
@@ -49,8 +49,8 @@ pub enum ValidityError {
     DuplicateInputUtxoId {
         utxo_id: UtxoId,
     },
-    DuplicateMessageInputId {
-        message_id: MessageId,
+    DuplicateInputNonce {
+        nonce: Nonce,
     },
     DuplicateInputContractId {
         contract_id: ContractId,